+++ test/cassettes/users/getNextPageOfChildren.yml
@@ -2,7 +2,7 @@
 -
     request:
         method: GET
-        url: 'https://api.easypost.com/beta/users/children?page_size=5'
+        url: 'https://api.easypost.com/v2/users/children?page_size=5'
         headers:
             Host: api.easypost.com
             Accept-Encoding: ''
@@ -21,58 +21,57 @@
             x-download-options: noopen
             x-permitted-cross-domain-policies: none
             referrer-policy: strict-origin-when-cross-origin
-            x-ep-request-uuid: dd5123f5657cb2dee78a0fa7001f9ad8
+            x-ep-request-uuid: 3c78c40b6597056bf0221b62000d5ffb
             cache-control: 'private, no-cache, no-store'
             pragma: no-cache
             expires: '0'
             content-type: 'application/json; charset=utf-8'
             content-length: '887'
-            x-runtime: '0.036851'
-            vary: Origin
-            x-node: bigweb41nuq
-            x-version-label: easypost-202312142151-39776a9e02-master
+            x-runtime: '0.035511'
+            x-node: bigweb36nuq
+            x-version-label: easypost-202401041812-437974c716-master
             x-backend: easypost
-            x-proxied: ['intlb2nuq b3de2c47ef', 'extlb2nuq 003ad9bca0']
+            x-proxied: ['intlb1nuq 2c48984abf', 'extlb2nuq 003ad9bca0']
             strict-transport-security: 'max-age=31536000; includeSubDomains; preload'
         body: '{"children":[{"id":"user_3c44214dcf744d169c3f2db034fd965e","object":"User","parent_id":"user_465eab8bb86844108100be2d4d5288ab","name":"Test User","phone_number":"<REDACTED>","verified":true,"created_at":"2022-11-23T21:03:03Z"},{"id":"user_78019b89199646358298fd311435458f","object":"User","parent_id":"user_465eab8bb86844108100be2d4d5288ab","name":"Test User","phone_number":"<REDACTED>","verified":true,"created_at":"2022-12-02T22:41:09Z"},{"id":"user_1a2e9fb7903c4a08b2da7d7f57150370","object":"User","parent_id":"user_465eab8bb86844108100be2d4d5288ab","name":"Test User","phone_number":"<REDACTED>","verified":true,"created_at":"2022-12-02T22:43:07Z"},{"id":"user_f01890d357374ac3837ff556a957b9b4","object":"User","parent_id":"user_465eab8bb86844108100be2d4d5288ab","name":"Test User","phone_number":"<REDACTED>","verified":true,"created_at":"2022-12-05T17:00:14Z"}],"has_more":false}'
         curl_info:
-            url: 'https://api.easypost.com/beta/users/children?page_size=5'
+            url: 'https://api.easypost.com/v2/users/children?page_size=5'
             content_type: 'application/json; charset=utf-8'
             http_code: 200
-            header_size: 702
-            request_size: 305
+            header_size: 688
+            request_size: 303
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.268365
-            namelookup_time: 0.029512
-            connect_time: 0.101179
-            pretransfer_time: 0.169023
+            total_time: 0.15691
+            namelookup_time: 0.001169
+            connect_time: 0.042299
+            pretransfer_time: 0.081692
             size_upload: 0.0
             size_download: 887.0
-            speed_download: 3305.0
+            speed_download: 5652.0
             speed_upload: 0.0
             download_content_length: 887.0
             upload_content_length: 0.0
-            starttransfer_time: 0.26829
+            starttransfer_time: 0.156866
             redirect_time: 0.0
             redirect_url: ''
             primary_ip: 169.62.110.130
             certinfo: {  }
             primary_port: 443
-            local_ip: 10.130.6.20
-            local_port: 56861
+            local_ip: 172.168.100.149
+            local_port: 57059
             http_version: 2
             protocol: 2
             ssl_verifyresult: 0
             scheme: HTTPS
-            appconnect_time_us: 168940
-            connect_time_us: 101179
-            namelookup_time_us: 29512
-            pretransfer_time_us: 169023
+            appconnect_time_us: 81617
+            connect_time_us: 42299
+            namelookup_time_us: 1169
+            pretransfer_time_us: 81692
             redirect_time_us: 0
-            starttransfer_time_us: 268290
-            total_time_us: 268365
+            starttransfer_time_us: 156866
+            total_time_us: 156910
             effective_method: GET
             capath: ''
             cainfo: ''
